+++ src/main/java/com/atlauncher/gui/tabs/PacksTab.java
@@ -41,6 +41,7 @@ import com.atlauncher.evnt.manager.RelocalizationManager;
 import com.atlauncher.evnt.manager.TabChangeManager;
 import com.atlauncher.gui.card.NilCard;
 import com.atlauncher.gui.card.PackCard;
+import com.atlauncher.gui.dialogs.AddCursePackDialog;
 import com.atlauncher.gui.dialogs.AddPackDialog;
 import com.atlauncher.gui.panels.LoadingPanel;
 import com.atlauncher.network.Analytics;
@@ -53,6 +54,7 @@ public final class PacksTab extends JPanel implements Tab, RelocalizationListene
     private final JPanel contentPanel = new JPanel(new GridBagLayout());
     private final JPanel bottomPanel = new JPanel(new FlowLayout(FlowLayout.CENTER));
     private final JButton addButton = new JButton(GetText.tr("Add Pack"));
+    private final JButton addCurseButton = new JButton(GetText.tr("Add Curse Pack"));
     private final JButton clearButton = new JButton(GetText.tr("Clear"));
     private final JButton expandAllButton = new JButton(GetText.tr("Expand All"));
     private final JButton collapseAllButton = new JButton(GetText.tr("Collapse All"));
@@ -118,6 +120,9 @@ public final class PacksTab extends JPanel implements Tab, RelocalizationListene
             new AddPackDialog();
             reload();
         });
+        this.addCurseButton.addActionListener(e -> {
+            new AddCursePackDialog();
+        });
         this.clearButton.addActionListener(e -> {
             searchField.setText("");
             searchDescBox.setSelected(false);
@@ -156,6 +161,7 @@ public final class PacksTab extends JPanel implements Tab, RelocalizationListene
 
     private void setupTopPanel() {
         this.topPanel.add(this.addButton);
+        this.topPanel.add(this.addCurseButton);
         this.topPanel.add(this.clearButton);
         this.topPanel.add(this.searchField);
         this.topPanel.add(this.searchButton);
@@ -271,6 +277,7 @@ public final class PacksTab extends JPanel implements Tab, RelocalizationListene
     @Override
     public void onRelocalization() {
         addButton.setText(GetText.tr("Add Pack"));
+        addCurseButton.setText(GetText.tr("Add Curse Pack"));
         clearButton.setText(GetText.tr("Clear"));
         expandAllButton.setText(GetText.tr("Expand All"));
         collapseAllButton.setText(GetText.tr("Collapse All"));