+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 3.3.1.0
 
 - Add in servers tab to manage servers created through the launcher
-- Add Curse packs by dragging the pack zip in instances tab
+- Add Curse packs from the packs tab